requirements.txt

flask==3.0.0
twilio==9.0.0
gspread==5.12.0
python-dotenv==1.0.0
gunicorn==21.2.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0